requirements.txt
alembic==1.7.7
fastapi==0.75.2
Jinja2==3.1.2
psycopg2-binary==2.9.3
PyJWT==2.3.0
python-decouple==3.6
python-dotenv==0.20.0
SQLAlchemy==1.4.36
uvicorn==0.17.6
python-multipart==0.0.5
python-jose==3.3.0
passlib==1.7.4
